Fixed and improved the sidepanes
Put the code of left and right sidepanels into two different files, remove the buttons text1-text8, remove the feature that clicking on a blank page will exit the sidepanels, making the movement of the right sidepanel smoother
